+++ src/hw-1/gui_components/filterspanel.c
@@ -230,7 +230,7 @@ void on_edges_btn_click(GtkWidget *caller
     GdkPixbuf *src_pixbuf = ((ImageFileData *)data)->pixbuf;
     const gchar *filename = ((ImageFileData *)data)->filename;
 
-    float *kernel = &edges_kernel_5x5[0][0];
+    double *kernel = &edges_kernel_5x5[0][0];
     gint    ker_width = 5, ker_height = 5;
     double  factor = 1.00, bias = 0.0;
